server/pyproject.toml
[project]
name = "hanpipette-chat-server"
version = "0.1.0"
description = "FastAPI chat backend for hanpipette.uk"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "bcrypt>=4.2",
    "pyjwt>=2.10",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false
